// frontend/src/api/generated/model/toolSettingFieldMetaType.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * TradingAgents Web API
 * AI-powered trading dashboard with simulation and live trading support
 * OpenAPI spec version: 1.0.0
 */

export type ToolSettingFieldMetaType = typeof ToolSettingFieldMetaType[keyof typeof ToolSettingFieldMetaType];


export const ToolSettingFieldMetaType = {
  boolean: 'boolean',
  number: 'number',
  string: 'string',
  textarea: 'textarea',
  select: 'select',
  multi_select: 'multi_select',
  string_list: 'string_list',
  secret: 'secret',
} as const;